+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/GameObject/AddComponentHandler.cs
@@ -113,7 +113,7 @@ namespace UniCli.Server.Editor.Handlers
     [Serializable]
     public class AddComponentRequest
     {
-        public int instanceId;
+        public long instanceId;
         public string path = "";
         public string typeName;
     }
@@ -123,7 +123,7 @@ namespace UniCli.Server.Editor.Handlers
     {
         public string gameObjectName;
         public string typeName;
-        public int instanceId;
+        public long instanceId;
         public bool enabled;
     }
 }
