+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/GameObject/CreateGameObjectHandler.cs
@@ -144,7 +144,7 @@ namespace UniCli.Server.Editor.Handlers
     [Serializable]
     public class CreateGameObjectResponse
     {
-        public int instanceId;
+        public long instanceId;
         public string name;
         public string path;
         public bool isActive;
