+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/GameObject/DestroyGameObjectHandler.cs
@@ -47,7 +47,7 @@ namespace UniCli.Server.Editor.Handlers
     [Serializable]
     public class DestroyGameObjectRequest
     {
-        public int instanceId;
+        public long instanceId;
         public string path = "";
     }
 
@@ -55,6 +55,6 @@ namespace UniCli.Server.Editor.Handlers
     public class DestroyGameObjectResponse
     {
         public string name;
-        public int instanceId;
+        public long instanceId;
     }
 }
