+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/Prefab/PrefabGetStatusHandler.cs
@@ -68,7 +68,7 @@ namespace UniCli.Server.Editor.Handlers
     [Serializable]
     public class PrefabGetStatusRequest
     {
-        public int instanceId;
+        public long instanceId;
         public string path = "";
     }
