+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/Prefab/PrefabSaveHandler.cs
@@ -60,7 +60,7 @@ namespace UniCli.Server.Editor.Handlers
     [Serializable]
     public class PrefabSaveRequest
     {
-        public int instanceId;
+        public long instanceId;
         public string path = "";
         public string assetPath;
     }
